+++ test/index.js
@@ -1,12 +1,15 @@
 /* global describe:true,it:true, expect: true */
 /* jshint expr: true */
 
-var proxy = require('proxy-call'),
-    expect = require('chai').expect;
+var Promise = require('promise'),
+    expect = require('chai').expect,
+    proxy = (typeof window === 'undefined')
+      ? require('..')
+      : require('proxy-call');
 
 describe('Proxy specs', function() {
 
-  it('should proxy method (success)', function(done) {
+  it('should proxy method (fulfilled promise)', function(done) {
 
     var obj = {
       foo: function(arg) {
@@ -19,20 +22,18 @@ describe('Proxy specs', function() {
     function load() {
       return new Promise(function(resolve) {
         setTimeout(function() {
-
           obj.loaded = true;
           resolve();
         }, 10);
       });
     }
 
-    proxy(obj, ['foo'], load);
-    obj
+    proxy(obj, ['foo'], load)
       .foo('bar')
       .then(function() { done() });
   });
 
-  it('should proxy method (failure)', function(done) {
+  it('should proxy method (rejected promise)', function(done) {
 
     var obj = {
       foo: function() {
@@ -46,14 +47,36 @@ describe('Proxy specs', function() {
       });
     }
 
-    proxy(obj, ['foo'], load);
-
-    obj
+    proxy(obj, ['foo'], load)
       .foo('bar')
       .catch(function(err) {
         expect(err).to.eq('booum');
         done();
       });
+  });
+
+  it('should proxy all methods when argument length === 2', function(done) {
+    var obj = {
+      bar: 'bar',
+      foo: function(arg) {
+        expect(this).to.be.eq(obj);
+        expect(this.loaded).to.be.ok;
+        expect(arg).to.be.eq('bar');
+      }
+    };
+
+    function load() {
+      return new Promise(function(resolve) {
+        setTimeout(function() {
+          obj.loaded = true;
+          resolve();
+        }, 10);
+      });
+    }
+
+    proxy(obj, load)
+      .foo('bar')
+      .then(function() { done() });
   });
 
 });