+++ HISTORY.md
@@ -1,3 +1,6 @@
-# 1.0.0 / 2013-01-01
 
-  * Initial release
+0.0.2 / 2014-06-17
+==================
+
+  * return proxy object
+  * make tests work on node/browsers